requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
autopep8==1.5.1
chardet==3.0.4
idna==2.9
multidict==4.7.5
pycodestyle==2.5.0
slackclient==2.5.0
tinydb==3.15.2
yarl==1.4.2
